added mox to c2b test
